scheduler: replace audit logic with timeout context
